refactor: restructure main orders component to complete modular architecture
BREAKING CHANGE: Convert final orders component from inline template to modular structure

Changes:
- Refactor main orders component into separate files (HTML, CSS, TS, index)
- Update route import to use new modular component path
- Remove old monolithic orders component file
- Maintain all existing functionality and OrderStatsComponent integration

Files:
- apps/frontend/src/app/private/modules/store/orders/orders/
- apps/frontend/src/app/routes/private/store_admin.routes.ts

Complete modular architecture achieved:
- ✅ order-stats (previously refactored)
- ✅ orders-list (previously refactored)
- ✅ order-details (previously refactored)
- ✅ orders (newly refactored)

All components now follow consistent modular pattern:
- Separation of concerns (HTML, CSS, TypeScript)
- Improved maintainability and developer experience
- Consistent export patterns with index.ts files
- Better caching and bundle optimization
