+++ apps/frontend/src/app/private/modules/store/orders/orders/index.ts
@@ -1,0 +1,1 @@
+export { OrdersComponent } from './orders.component';
