+++ apps/frontend/src/app/private/modules/store/orders/orders/orders.component.css
@@ -1,0 +1,4 @@
+:host {
+  display: block;
+  width: 100%;
+}
