+++ apps/frontend/src/app/private/modules/store/orders/orders/orders.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { OrderStatsComponent } from '../components/order-stats';
+
+@Component({
+  selector: 'app-orders',
+  standalone: true,
+  imports: [CommonModule, OrderStatsComponent],
+  templateUrl: './orders.component.html',
+  styleUrls: ['./orders.component.css'],
+})
+export class OrdersComponent {}
